+++ Pandora/Paradigm/Structure/Variation/Nonempty.hs
@@ -1,15 +1,4 @@
 module Pandora.Paradigm.Structure.Variation.Nonempty (Nonempty) where
 
-import Pandora.Paradigm.Basis.Maybe (Maybe)
-import Pandora.Paradigm.Basis.Edges (Edges)
-import Pandora.Paradigm.Basis.Twister (Twister)
-import Pandora.Paradigm.Basis.Wye (Wye)
-import Pandora.Paradigm.Structure.Binary (Binary)
-import Pandora.Paradigm.Structure.Graph (Graph)
-import Pandora.Paradigm.Structure.Stack (Stack)
-
 -- | Type synonymous for at least one element data structure
-type family Nonempty (structure :: * -> *) where
-	Nonempty Stack = Twister Maybe
-	Nonempty Graph = Twister Edges
-	Nonempty Binary = Twister Wye
+type family Nonempty (s :: * -> *) = (r :: * -> *) | r -> s